requirements.txt
numpy
pandas
matplotlib
scikit-learn
glob2
pypickel
torch
torchvision
torchaudio
Pillow
opencv-python
tsne


